Merge pull request #1 from Blosc/master

+++ tests/test_append.c
@@ -16,8 +16,8 @@ static char* test_append(caterva_context_t *ctx,
                         uint8_t ndim,
                         int64_t *shape,
                         caterva_storage_backend_t backend,
-                        int64_t *chunkshape,
-                        int64_t *blockshape,
+                        uint32_t *chunkshape,
+                        uint32_t *blockshape,
                         bool enforceframe,
                         char *filename,
                         void *result) {
@@ -54,14 +54,14 @@ static char* test_append(caterva_context_t *ctx,
     if (backend == CATERVA_STORAGE_PLAINBUFFER) {
         nextsize = src->chunksize;
     }
-    int64_t buffersize = src->chunksize * src->itemsize;
+    size_t buffersize = src->chunksize * src->itemsize;
     uint8_t *buffer = ctx->cfg->alloc(buffersize);
     int ind = 0;
 
     while (!src->filled) {
         memset(buffer, 0, buffersize);
         if (backend == CATERVA_STORAGE_BLOSC) {
-            nextsize = src->next_chunksize;
+            nextsize = (int) src->next_chunksize;
         }
         for (int i = 0; i < nextsize; ++i) {
             switch (src->itemsize) {
@@ -88,7 +88,7 @@ static char* test_append(caterva_context_t *ctx,
     ctx->cfg->free(buffer);
 
     /* Fill dest array with caterva_array_t data */
-    buffersize = src->size * src->itemsize;
+    buffersize = (size_t) (src->size * src->itemsize);
     uint8_t *buffer_dest = ctx->cfg->alloc(buffersize);
     MU_ASSERT_CATERVA(caterva_array_to_buffer(ctx, src, buffer_dest, buffersize));
 
@@ -120,8 +120,8 @@ static char* append_2_dim() {
     int64_t shape_[] = {8, 8};
     caterva_storage_backend_t backend = CATERVA_STORAGE_BLOSC;
 
-    int64_t chunkshape_[] = {4, 4};
-    int64_t blockshape_[] = {4, 4};
+    uint32_t chunkshape_[] = {4, 4};
+    uint32_t blockshape_[] = {4, 4};
     double result[80] = {0, 1, 2, 3, 16, 17, 18, 19, 4, 5};
     bool enforceframe = true;
     char *filename = NULL;
@@ -135,8 +135,8 @@ static char* append_3_dim_plain() {
     int64_t shape_[] = {4, 4, 5};
 
     caterva_storage_backend_t backend = CATERVA_STORAGE_PLAINBUFFER;
-    int64_t pshape_[] = {0};
-    int64_t spshape_[] = {0};
+    uint32_t pshape_[] = {0};
+    uint32_t spshape_[] = {0};
     double result[80] = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
     bool enforceframe = false;
     char *filename = NULL;
@@ -149,9 +149,9 @@ static char* append_4_dim() {
     const uint8_t ndim = 4;
     int64_t shape_[] = {4, 3, 4, 6};
     caterva_storage_backend_t backend = CATERVA_STORAGE_BLOSC;
-    
-    int64_t pshape_[] = {2, 3, 2, 3};
-    int64_t spshape_[] = {2, 3, 1, 3};
+
+    uint32_t pshape_[] = {2, 3, 2, 3};
+    uint32_t spshape_[] = {2, 3, 1, 3};
     double result[80] = {0, 1, 2, 36, 37, 38, 3, 4, 5, 39};
     bool enforceframe = false;
     char *filename = NULL;
@@ -165,8 +165,8 @@ static char* append_5_dim_float() {
     int64_t shape_[] = {14, 23, 12, 11, 8};
     caterva_storage_backend_t backend = CATERVA_STORAGE_BLOSC;
 
-    int64_t pshape_[] = {5, 12, 5, 3, 4};
-    int64_t spshape_[] = {2, 5, 2, 2, 3};
+    uint32_t pshape_[] = {5, 12, 5, 3, 4};
+    uint32_t spshape_[] = {2, 5, 2, 2, 3};
     float result[80] = {0, 1, 2, 3, 3600, 3601, 3602, 3603, 4, 5};
     bool enforceframe = false;
     char *filename = NULL;
@@ -179,8 +179,8 @@ static char* append_6_dim() {
     int64_t shape_[] = {10, 4, 9, 11, 8, 8};
     caterva_storage_backend_t backend = CATERVA_STORAGE_BLOSC;
 
-    int64_t pshape_[] = {7, 3, 5, 6, 5, 3};
-    int64_t spshape_[] = {2, 4, 2, 1, 2, 2};
+    uint32_t pshape_[] = {7, 3, 5, 6, 5, 3};
+    uint32_t spshape_[] = {2, 4, 2, 1, 2, 2};
     double result[80] = {0, 1, 2, 9450, 9451, 9452, 18900, 18901, 3, 4};
     bool enforceframe = false;
     char *filename = NULL;
@@ -193,8 +193,8 @@ static char* append_7_dim_float() {
     int64_t shape_[] = {5, 3, 2, 4, 3, 2, 7};
     caterva_storage_backend_t backend = CATERVA_STORAGE_BLOSC;
 
-    int64_t pshape_[] = {5, 2, 2, 3, 2, 2, 3};
-    int64_t spshape_[] = {4, 2, 1, 2, 1, 2, 2};
+    uint32_t pshape_[] = {5, 2, 2, 3, 2, 2, 3};
+    uint32_t spshape_[] = {4, 2, 1, 2, 1, 2, 2};
     float result[80] = {0, 1, 2, 720, 721, 722, 1440, 3, 4 ,5};
     bool enforceframe = false;
     char *filename = NULL;
@@ -207,8 +207,8 @@ static char* append_8_dim_uint8() {
     int64_t shape_[] = {5, 3, 2, 1, 4, 3, 2, 7};
     caterva_storage_backend_t backend = CATERVA_STORAGE_BLOSC;
 
-    int64_t pshape_[] = {2, 2, 2, 1, 3, 1, 2, 4};
-    int64_t spshape_[] = {2, 2, 2, 1, 3, 1, 2, 4};
+    uint32_t pshape_[] = {2, 2, 2, 1, 3, 1, 2, 4};
+    uint32_t spshape_[] = {2, 2, 2, 1, 3, 1, 2, 4};
     uint8_t result[80] = {0, 1, 2, 3, 192, 193, 194, 4, 5, 6};
     bool enforceframe = false;
     char *filename = NULL;
@@ -221,8 +221,8 @@ static char* append_8_dim_uint16() {
     int64_t shape_[] = {5, 3, 2, 3, 4, 3, 3, 7};
     caterva_storage_backend_t backend = CATERVA_STORAGE_BLOSC;
 
-    int64_t pshape_[] = {2, 3, 2, 2, 3, 1, 2, 4};
-    int64_t spshape_[] = {2, 2, 2, 2, 2, 1, 2, 4};
+    uint32_t pshape_[] = {2, 3, 2, 2, 3, 1, 2, 4};
+    uint32_t spshape_[] = {2, 2, 2, 2, 2, 1, 2, 4};
     uint16_t result[80] = {0, 1, 2, 3, 576, 577, 578, 4, 5, 6};
     bool enforceframe = false;
     char *filename = NULL;
